+++ web/src/pages/SingleQuestionPage.js
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react'
 import { connect } from 'react-redux'
 import { deleteAnswer } from '../actions/questionActions'
 import { fetchQuestion } from '../actions/questionActions'
-
+import swal from 'sweetalert';
 import { Question } from '../components/Question'
 import { Answer } from '../components/Answer'
 import { Link } from 'react-router-dom'
@@ -13,12 +13,13 @@ const SingleQuestionPage = ({
   question,
   hasErrors,
   loading,
-  userId
+  userId,
+  redirect
 }) => {
   const { id } = match.params
   useEffect(() => {
     dispatch(fetchQuestion(id))
-  }, [dispatch, id])
+  }, [dispatch, id, redirect])
 
   const renderQuestion = () => {
     if (loading.question) return <p>Loading question...</p>
@@ -28,7 +29,22 @@ const SingleQuestionPage = ({
   }
 
   const onDeleteAnswer = (answer) => {
-    dispatch(deleteAnswer(answer))
+    swal({
+      title: "Are you sure?",
+      text: "Once deleted, you will not be able to recover this answer!",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    })
+    .then((willDelete) => {
+      if (willDelete) {
+        dispatch(deleteAnswer(answer))
+        swal("Poof! Your answer has been deleted!", {
+          icon: "success",
+        });
+      }
+    });
+    
 }
 
   const renderAnswers = () => {
@@ -54,7 +70,8 @@ const mapStateToProps = state => ({
   question: state.question.question,
   loading: state.question.loading,
   hasErrors: state.question.hasErrors,
-  userId: state.auth.uid
+  userId: state.auth.uid,
+  redirect: state.question.redirect
 })
 
 export default connect(mapStateToProps)(SingleQuestionPage)
